src/requirements.txt
pandas
sqlalchemy
psycopg2-binary
python-dotenv
yfinance
